bin/build*: Check that required binaries are present

+++ bin/build.sh
@@ -1,5 +1,27 @@
 #!/usr/bin/env bash
 
+set -e
+
+if ! command -v jq &> /dev/null; then
+  echo "Missing command: jq"
+  exit 1
+fi
+
+if ! command -v fakeroot &> /dev/null; then
+  echo "Missing command: fakeroot"
+  exit 1
+fi
+
+if ! command -v dpkg-deb &> /dev/null; then
+  echo "Missing command: dpkg-deb"
+  exit 1
+fi
+
+if ! command -v wine &> /dev/null; then
+  echo "Missing command: wine"
+  exit 1
+fi
+
 echo "Deleting existing builds"
 
 rm -rf dist/*
